+++ app/charts/area/areas-state.tsx
@@ -48,6 +48,10 @@ import {
 } from "@/charts/shared/interaction/tooltip-box";
 import { DEFAULT_MARGIN_TOP } from "@/charts/shared/margins";
 import {
+  ShowTemporalValueLabelsVariables,
+  useShowTemporalValueLabelsVariables,
+} from "@/charts/shared/show-values-utils";
+import {
   getStackedTooltipValueFormatter,
   getStackedYScale,
 } from "@/charts/shared/stacked-helpers";
@@ -70,7 +74,8 @@ import { ChartProps } from "../shared/ChartProps";
 
 export type AreasState = CommonChartState &
   AreasStateVariables &
-  InteractiveXTimeRangeState & {
+  InteractiveXTimeRangeState &
+  Omit<ShowTemporalValueLabelsVariables, "yOffset"> & {
     chartType: "area";
     xScale: ScaleTime<number, number>;
     yScale: ScaleLinear<number, number>;
@@ -89,7 +94,7 @@ const useAreasState = (
   variables: AreasStateVariables,
   data: ChartStateData
 ): AreasState => {
-  const { chartConfig } = chartProps;
+  const { chartConfig, dimensions, measures } = chartProps;
   const {
     xDimension,
     getX,
@@ -114,6 +119,7 @@ const useAreasState = (
     allData,
   } = data;
   const { fields, interactiveFiltersConfig } = chartConfig;
+  const { y } = fields;
 
   const { width, height } = useSize();
   const formatNumber = useFormatNumber({ decimals: "auto" });
@@ -349,6 +355,11 @@ const useAreasState = (
     normalize,
   });
   const right = 40;
+
+  const chartWidth = getChartWidth({ width, left, right });
+  xScale.range([0, chartWidth]);
+  xScaleTimeRange.range([0, chartWidth]);
+
   const leftAxisLabelSize = useAxisLabelSizeVariables({
     label: yAxisLabel,
     width,
@@ -361,19 +372,21 @@ const useAreasState = (
     marginLeft: left,
     marginRight: right,
   });
+  const { yOffset: yValueLabelsOffset, ...showValuesVariables } =
+    useShowTemporalValueLabelsVariables(y, {
+      dimensions,
+      measures,
+      segment: fields.segment,
+    });
   const margins = {
-    top: DEFAULT_MARGIN_TOP + leftAxisLabelSize.offset,
+    top: DEFAULT_MARGIN_TOP + leftAxisLabelSize.offset + yValueLabelsOffset,
     right,
     bottom,
     left,
   };
-  const chartWidth = getChartWidth({ width, left, right });
   const bounds = useChartBounds({ width, chartWidth, height, margins });
   const { chartHeight } = bounds;
 
-  /** Adjust scales according to dimensions */
-  xScale.range([0, chartWidth]);
-  xScaleTimeRange.range([0, chartWidth]);
   yScale.range([chartHeight, 0]);
 
   const isMobile = useIsMobile();
@@ -471,6 +484,7 @@ const useAreasState = (
     getAnnotationInfo,
     leftAxisLabelSize,
     bottomAxisLabelSize,
+    ...showValuesVariables,
     ...variables,
   };
 };